Refactor: move shells into xonsh.shell
