+++ tests/test_ptk_history.py
@@ -9,7 +9,7 @@ except ImportError:
 @pytest.fixture
 def history_obj():
     """Instantiate `PromptToolkitHistory` and append a line string"""
-    from xonsh.ptk_shell.history import PromptToolkitHistory
+    from xonsh.shell.ptk.history import PromptToolkitHistory
 
     hist = PromptToolkitHistory(load_prev=False)
     hist.append_string("line10")
@@ -27,7 +27,7 @@ def test_ptk2_backcompat():
     Test that legacy code (ahem, xontribs) can still reference xonsh.ptk2 (for a while)
     """
 
-    import xonsh.ptk_shell.shell as imports_new
+    import xonsh.shell.ptk.shell as imports_new
     import xonsh.ptk2.shell as imports_legacy
 
     # defining the ptk2 package this way leaves out the internal global names (which all start with '_')
