+++ tests/test_readline_shell.py
@@ -1,7 +1,7 @@
 import pytest
 
 from xonsh.completers.tools import RichCompletion
-from xonsh.readline_shell import _render_completions
+from xonsh.shell.readline import _render_completions
 
 
 @pytest.mark.parametrize(
