+++ xonsh/ptk2/key_bindings.py
@@ -1,1 +1,1 @@
-from xonsh.ptk_shell.key_bindings import *  # noqa: F403 F401
+from xonsh.shell.ptk import *  # noqa: F403 F401
