+++ xonsh/shell/base.py
