+++ xonsh/shell/jupyter.py
@@ -2,7 +2,7 @@
 import io
 import sys
 
-from xonsh.base_shell import BaseShell
+from xonsh.shell.base import BaseShell
 from xonsh.built_ins import XSH
