+++ xonsh/shell/ptk/completer.py
